Added card pre-render test and code refactor
