+++ packages/realm-server/tests/helpers/index.ts
@@ -1,7 +1,8 @@
 import { writeFileSync, writeJSONSync } from "fs-extra";
 import { NodeAdapter } from "../../node-realm";
-import { join } from "path";
+import { resolve, join } from "path";
 import { Realm, LooseSingleCardDocument } from "@cardstack/runtime-common";
+import { makeFastBoot, visit } from "../../fastboot";
 import type * as CardAPI from "https://cardstack.com/base/card-api";
 
 export const testRealm = "http://test-realm/";
@@ -18,9 +19,12 @@ export function createRealm(
       writeJSONSync(join(dir, filename), contents);
     }
   }
-  return new Realm(realmURL, new NodeAdapter(dir), async () => {
-    return `card render is not implemented`;
-  });
+  return new Realm(
+    realmURL,
+    new NodeAdapter(dir),
+    visit,
+    makeFastBoot(resolve(__dirname, "..", "..", "..", "host", "dist"))
+  );
 }
 
 export function setupCardLogs(
